+++ fs/proc/pid.c
@@ -111,12 +111,12 @@ static ssize_t proc_pid_cmdline_show(struct proc_entry *entry, char *buf) {
 
 static struct proc_dir_entry proc_pid_fd;
 
-static bool proc_pid_fd_readdir(struct proc_entry *entry, int *index, struct proc_entry *next_entry) {
+static bool proc_pid_fd_readdir(struct proc_entry *entry, unsigned *index, struct proc_entry *next_entry) {
     struct task *task = proc_get_task(entry);
     if (task == NULL)
         return _ESRCH;
     lock(&task->files->lock);
-    while (*index < task->files->size && task->files->files[*index] == NULL)
+    while ((int) *index < task->files->size && task->files->files[*index] == NULL)
         (*index)++;
     fd_t f = (*index)++;
     bool any_left = f < task->files->size;
